requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
starlette==0.37.2
pydantic==2.9.2
python-dotenv==1.0.1
pydantic[email]==2.9.2
pymongo[srv]==4.13.2
passlib==1.7.4
bcrypt==3.2.2  
python-jose==3.3.0
